Make job synchron and move it to shared angular jobs
