+++ packages/components/scripts/jobs/angular/index.js
@@ -4,3 +4,4 @@ export * from './createFormsModule.js';
 export * from './createNgModule.js';
 export * from './createValidatorDirectives.js';
 export * from './build.js';
+export * from './shared.js';
